requirement.txt
Django==5.2.7                    # The main framework
djangorestframework==3.14.0      # For creating APIs
djangorestframework-simplejwt==5.3.1  # For login tokens
django-cors-headers==4.3.1       # Allows frontend to talk to backend
django-allauth==0.58.2           # For Google login
whitenoise==6.6.0                # Handles CSS/JS files
gunicorn==21.2.0                 # Runs your app on server
dj-database-url==2.1.0           # Database connection helper
psycopg2-binary==2.9.9           # For PostgreSQL database
requests==2.31.0                 # For making HTTP requests
coverage==7.4.0